+++ src/models/mod.rs
@@ -1,11 +1,7 @@
-pub mod authentication_session;
-pub mod certificate_choice_session;
+pub mod api;
 pub mod common;
 pub mod device_link;
 pub mod interaction;
-mod response;
 pub mod semantic_identifier;
-pub mod session_status;
 pub mod signature;
-pub mod signature_session;
 pub mod user_identity;
